refactor: move logic out of DraggableElement component
